remove dependencies propagating

+++ src/OrleansRuntime/Scheduler/DownstreamContext.cs
@@ -7,11 +7,11 @@ namespace Orleans.Runtime.Scheduler
     [Serializable]
     public class DownstreamContext
     {
-        public Dictionary<string, double> StatsUpdate { get; }
+        public Dictionary<string, long> StatsUpdate { get; }
 
         public long MaximumDownstreamCost { get; }
 
-        public DownstreamContext(Dictionary<string, double> statsCollection, long maximumDownstreamCost = SchedulerConstants.DEFAULT_WIG_EXECUTION_COST)
+        public DownstreamContext(Dictionary<string, long> statsCollection, long maximumDownstreamCost = SchedulerConstants.DEFAULT_WIG_EXECUTION_COST)
         {
             StatsUpdate = statsCollection;
             MaximumDownstreamCost = maximumDownstreamCost;
